+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Version 7.1.2beta3
 ## Gameplay
 * Training job less likely to select skills close to skill cap
+* New Tattoo/Piercing management in office interactions
 
 ## Fixes
 * Set correct matron job for brothel. This fixes obedience calculations.